add RelayNetworkTypes
